Generate PDFs using Playwright
This commit adds a few new things:
- A SQS queue for background tasks
- A worker container that reads from the queue and executes the tasks
- A UI container that runs the UI within the docker network that's used
  Playwright to generate PDFs
- A utility function to generate PDFs using Playwright

The base image for the API is now based on the official Playwright
image. The image is reused for the worker container as well.

The pdf download API endpoint has been changed to queue a task for
generating the PDF through Playwright when a request is received for
downloading a "Document" PDF.